Fixes for Aerh.
