+++ src/main/java/net/hypixel/nerdbot/util/watcher/rss/SkyblockUpdateDataHandler.java
@@ -22,17 +22,13 @@ public class SkyblockUpdateDataHandler {
             return;
         }
 
-
         MessageCreateBuilder messageCreateBuilder = new MessageCreateBuilder();
-
         PingableRole role = RoleManager.getPingableRoleByName("SkyBlock Update Alerts");
-
         if (role != null) {
             messageCreateBuilder.addContent(RoleManager.formatPingableRoleAsMention(role) + "\n\n");
         }
 
         messageCreateBuilder.addContent(hypixelThread.getLink());
-
         announcementChannel.sendMessage(messageCreateBuilder.build()).queue();
     }
 }
